requirements.txt
twill
webdriver-manager
selenium
python-dotenv